+++ octasine/src/common.rs
@@ -2,20 +2,20 @@
 
 /// Number that gets incremented with 1.0 every second
 #[derive(Debug, Copy, Clone, PartialEq)]
-pub struct TimeCounter(pub f32);
+pub struct TimeCounter(pub f64);
 
 /// Phase. value >= 0.0 && value < 1.0
 #[derive(Debug, Copy, Clone)]
-pub struct Phase(pub f32);
+pub struct Phase(pub f64);
 
 #[derive(Debug, Copy, Clone)]
-pub struct SampleRate(pub f32);
+pub struct SampleRate(pub f64);
 
 #[derive(Debug, Copy, Clone)]
-pub struct TimePerSample(pub f32);
+pub struct TimePerSample(pub f64);
 
 #[derive(Debug, Copy, Clone)]
-pub struct BeatsPerMinute(pub f32);
+pub struct BeatsPerMinute(pub f64);
 
 #[derive(Debug, Copy, Clone, PartialEq)]
 pub enum EnvelopeStage {
