+++ octasine/src/processing_parameters/common.rs
@@ -8,15 +8,15 @@ use crate::common::TimeCounter;
 pub trait ParameterValueConversion {
     type ProcessingParameterValue;
 
-    fn to_processing(value: f32) -> Self::ProcessingParameterValue;
-    fn to_preset(value: Self::ProcessingParameterValue) -> f32;
+    fn to_processing(value: f64) -> Self::ProcessingParameterValue;
+    fn to_preset(value: Self::ProcessingParameterValue) -> f64;
 
     /// Parse a string value coming from the host
     fn parse_string_value(value: String) -> Option<Self::ProcessingParameterValue>;
 
     fn format_processing(internal_value: Self::ProcessingParameterValue) -> String;
 
-    fn format_value(value: f32) -> String {
+    fn format_value(value: f64) -> String {
         Self::format_processing(Self::to_processing(value))
     }
 }
@@ -31,8 +31,8 @@ pub trait ProcessingParameter {
 }
 
 pub trait ProcessingParameterPresetValueAccess {
-    fn set_from_preset_value(&mut self, value: f32);
-    fn get_preset_target_value(&self) -> f32;
+    fn set_from_preset_value(&mut self, value: f64);
+    fn get_preset_target_value(&self) -> f64;
 }
 
 impl<P, T> ProcessingParameterPresetValueAccess for P
@@ -40,10 +40,10 @@ impl<P, T> ProcessingParameterPresetValueAccess for P
         ProcessingParameter<Value = T> +
         ParameterValueConversion<ProcessingParameterValue = T>
 {
-    fn set_from_preset_value(&mut self, value: f32){
+    fn set_from_preset_value(&mut self, value: f64){
         self.set_value(Self::to_processing(value));
     }
-    fn get_preset_target_value(&self) -> f32 {
+    fn get_preset_target_value(&self) -> f64 {
         Self::to_preset(self.get_target_value())
     }
 }
@@ -51,16 +51,16 @@ impl<P, T> ProcessingParameterPresetValueAccess for P
 
 #[derive(Debug, Copy, Clone)]
 pub struct InterpolatableProcessingValue {
-    pub target_value: f32,
-    current_value: f32,
-    step_size: f32,
+    pub target_value: f64,
+    current_value: f64,
+    step_size: f64,
     steps_remaining: u8,
     samples_remaining: u8,
     last_time: TimeCounter,
 }
 
 impl InterpolatableProcessingValue {
-    pub fn new(value: f32) -> Self {
+    pub fn new(value: f64) -> Self {
         Self {
             target_value: value,
             current_value: value,
@@ -72,11 +72,11 @@ impl InterpolatableProcessingValue {
     }
 
     /// Possibly advance interpolation and call callback, return value.
-    pub fn get_value<F: FnMut(f32)>(
+    pub fn get_value<F: FnMut(f64)>(
         &mut self,
         time: TimeCounter,
         callback_on_advance: &mut F
-    ) -> f32 {
+    ) -> f64 {
         if self.steps_remaining == 0 || INTERPOLATION_STEPS == 0 {
             return self.current_value;
         }
@@ -100,7 +100,7 @@ impl InterpolatableProcessingValue {
     }
 
     // Set target value, possibly restart interpolation
-    pub fn set_value(&mut self, value: f32){
+    pub fn set_value(&mut self, value: f64){
         self.target_value = value;
 
         if INTERPOLATION_STEPS == 0 {
